sync zpp with master

+++ ustring.cpp
@@ -138,7 +138,7 @@ PHP_METHOD(UString, __construct)
 {
 	char *value = NULL, 
 	     *codepage = NULL;
-	int vlen = 0, clen = 0;
+	size_t vlen = 0, clen = 0;
 
 	if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "s|s", &value, &vlen, &codepage, &clen) != SUCCESS) {
 		return;
@@ -371,8 +371,8 @@ PHP_METHOD(UString, getCodepage) {
 
 /* {{{ proto void UString::setDefaultCodepage(string codepage) */
 PHP_METHOD(UString, setDefaultCodepage) {
-	char *codepage = NULL;
-	int  clen = 0;
+	char   *codepage = NULL;
+	size_t clen = 0;
 
 	if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "s", &codepage, &clen) != SUCCESS) {
 		return;
@@ -540,8 +540,8 @@ PHP_MINFO_FUNCTION(ustring)
 
 /* {{{ proto UString u(string value) */
 PHP_FUNCTION(u) {
-    char *val;
-    int len;
+    char   *val;
+    size_t len;
 
     if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "s", &val, &len) != SUCCESS) {
         return;
